% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.gui.R
\name{calib.gui}
\alias{calib.gui}
\title{Run the graphical user interface for analysis of the data of lateral
flow assay (streptavidin assay)}
\usage{
calib.gui()
}
\description{
Run the graphical user interface for streptavidin assay
}
\examples{
calib.gui()
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
