% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.gui.R
\name{exp.gui}
\alias{exp.gui}
\title{Run the graphical user interface for analysis of the data of lateral
flow assay (sandwich assay)}
\usage{
\method{exp}{gui}()
}
\description{
Run the graphical user interface for sandwich assay
}
\examples{
exp.gui()
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
