% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Qdata-class}
\alias{Qdata-class}
\title{S4 Class Qdata to contain the data}
\description{
S4 Class Qdata to contain the data
}
\section{Slots}{

\describe{
\item{\code{initialData}}{contains the initial data, in the form of an object
of Class \code{data.frame}}

\item{\code{nsgData}}{contains the normalized intensity data of green quantum dots
(QDs) in the form of an object of Class \code{data.frame}}

\item{\code{nsrData}}{contains the normalized intensity data of red quantum dots
(QDs) in the form of an object of Class \code{data.frame}}

\item{\code{cgData}}{contains the combined replicates of normalized intensity of
green quantum dots in the form of an object of Class \code{data.frame}}

\item{\code{sdgData}}{contains the standard deviation of normalized intensity of
green QDs within the replicates in the form of an object of Class
\code{data.frame}}

\item{\code{crData}}{contains the combined replicates of normalized intensity of
red QDs in the form of an object of Class \code{data.frame}}

\item{\code{sdrData}}{contains the standard deviation of normalized intensity of
red QDs within the replicates in the form of an object of Class \code{data.frame}}

\item{\code{Concentration}}{contains the concentration data of calibration in the
form of an object of Class \code{data.frame}}

\item{\code{Concentration_CRP}}{contains the concentration data of CRP in the form
of an object of Class \code{data.frame}}

\item{\code{Concentration_IL6}}{contains the concentration data of IL6 in the form
of an object of Class \code{data.frame}}

\item{\code{CIData}}{contains the confidence interval data of combined replicates
in the form of an object of Class \code{data.frame}}

\item{\code{corData}}{contains the result of correlation of combined normalized
intensities of calibration with respect to their predicted values by
linear model in the form of an object of Class \code{data.frame}}

\item{\code{corData_exp}}{contains the result of correlation of combined normalized
intensities with respect to their predicted values by linear model in the
form of an object of Class \code{data.frame}}

\item{\code{lobData}}{contains the result of normalized intensities values of limit
of blank (LOB), limit of detection (LOD) and limit of quantification (LOQ) in
the form of an object of Class \code{data.frame}}
}}

