% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDCor_Cal.R
\docType{methods}
\name{QDCor_Cal}
\alias{QDCor_Cal}
\title{To compute the correlation of the combined replicates in case
of streptavidin assay}
\usage{
\S4method{QDCor_Cal}{Qdata}(Qdata)
}
\arguments{
\item{Qdata}{object of S4 Class \code{Qdata}, which is an output of function
\code{Repsd}.}
}
\value{
An object of S4 Class \code{Qdata}, with a new slot of corData to
save the results of correlation of combined replicates of green
and red QDs.
}
\description{
Function \code{QDCor_Cal} works on the output of the function
\code{Repsd}. Function \code{QDCor_Cal} computes the correlation of the combined
replicates of green and red QDs with respect to the predicted values
through their respective linear fitted model in case of the streptavidin
assay. The output of the function \code{QDCor_Cal} are saved in the slot
slot corData of the object of S4 Class \code{Qdata}.
}
\details{
Function \code{QDCor_Cal} works on the output of the function
\code{Repsd}. Function \code{QDCor_Cal} computes the correlation of the combined
replicates of green and red QDs with respect to the predicted values
through their respective linear fitted model in case of the streptavidin
assay. The output of the function \code{QDCor_Cal} are saved in the slot
slot corData of the object of S4 Class \code{Qdata}.
}
\examples{
#To combine the replicates of streptavidin assay
file_cal_1 <- system.file("exData", "QD_calibration_Qplex_new.csv", package = "ReadqPCR")
data_stp <- readQD(file = file_cal_1, type = ".csv", decp = ",")
data_rep_stp <- Repsd(data_stp)
#To compute the correlation
cor_stp <- QDCor_Cal(data_rep_stp)
# To visualize all the results
cor_stp
# To visualize only correlation results
slot(cor_stp,"corData")
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{Qdata}
