% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDCor_exp.R
\docType{methods}
\name{QDCor_exp}
\alias{QDCor_exp}
\title{To compute the correlation of the combined replicates in case
of sandwich assay}
\usage{
\S4method{QDCor_exp}{Qdata}(Qdata)
}
\arguments{
\item{Qdata}{object of S4 Class \code{Qdata}, which is an output of function
\code{Repsd}.}
}
\value{
An object of S4 Class \code{Qdata}, with a new slot of corData_exp to
save the results of correlation of combined replicates of green
and red QDs.
}
\description{
Function \code{QDCor_exp} works on the output of the function
\code{Repsd}. Function \code{QDCor_exp} computes the correlation of the combined
replicates of green and red QDs with respect to the predicted values
through their respective linear fitted model in case of the sandwich
assay. The output of the function \code{QDCor_exp} are saved in the slot
slot corData_exp of the object of S4 Class \code{Qdata}.
}
\details{
Function \code{QDCor_exp} works on the output of the function
\code{Repsd}. Function \code{QDCor_exp} computes the correlation of the combined
replicates of green and red QDs with respect to the predicted values
through their respective linear fitted model in case of the sandwich
assay. The output of the function \code{QDCor_exp} are saved in the slot
slot corData_exp of the object of S4 Class \code{Qdata}.
}
\examples{
##To read the first file (.csv)
file_il_d_1 <- system.file("exData", "ImageJ_CRPi_IL6d_Qplex.csv", package = "Qplex")
data1 <- readQD(file = file_il_d_1, decp = ".")
## to read the second file (.csv)
file_il_i_1 <- system.file("exData", "ImageJ_CRPi_IL6i_Qplex.csv", package = "Qplex")
data2 <- readQD(file = file_il_i_1, decp = ".")
data <- ReadQD_combine(data1, data2)
data_rep_sand <- Repsd(data, calib = FALSE)
# To compute the correlation
cor_sand <- QDCor_exp(data_rep_sand)
# to visualize all the results
cor_sand
# to visualize only the correlation results
slot(cor_sand,"corData_exp")
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{Qdata}
