% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDLOD_calib.R
\docType{methods}
\name{QDLOD_calib}
\alias{QDLOD_calib}
\title{To compute the LOB, LOD and LOQ of streptavidin assay}
\usage{
\S4method{QDLOD_calib}{Qdata}(Qdata, LOB = FALSE)
}
\arguments{
\item{Qdata}{object of S4 Class \code{Qdata}, which is an output of function
\code{Repsd}.}

\item{LOB}{Choose between TRUE and FALSE. FALSE is for computing the
LOD and LOQ intensities by first method, without involving the LOB. TRUE is for computing all
three LOB, LOD and LOQ intensities by second method, with involvement
of LOB in calculation of LOD too. Second method is more accurate method.}
}
\value{
An object of S4 Class \code{Qdata}, with a new slot of lobData to
save the results of LOB, LOD and LOQ of combined replicates of green
and red QDs.
}
\description{
Function \code{QDLOD_calib} works on the output of the function
\code{Repsd}. Function \code{QDLOD_calib} computes the limit of blank (LOB),
limit of detection (LOD) and limit of quantification (LOQ) intensities
of the streptavidin assay. These intensities can be used to compute
the LOB, LOD and LOQ concentration, based on the linear model equation.
}
\details{
Function \code{QDLOD_calib} works on the output of the function
\code{Repsd}. Function \code{QDLOD_calib} computes the limit of blank (LOB),
limit of detection (LOD) and limit of quantification (LOQ) intensities
of the streptavidin assay. The results of LOB, LOD and LOQ intensities are
saved in the new slot lobData of the object of S4 Class \code{Qdata}. These intensities can be used to compute
the LOB, LOD and LOQ concentration, based on the linear model equation.
}
\examples{
#To combine the replicates of streptavidin assay
file_cal_1 <- system.file("exData", "QD_calibration_Qplex_new.csv", package = "ReadqPCR")
data_stp <- readQD(file = file_cal_1, type = ".csv", decp = ",")
#to combine the replicates
data_rep_stp <- Repsd(data_stp)
# 1) To compute the LOB, LOD and LOQ intensities by
# first method
lod_stp1 <- QDLOD_calib(data_rep_stp)
#to visualize all the results
lod_stp1
#to visualize only the lob, lod and loq intensities
slot(lod_stp1,"lobData")
# 2) To compute the LOB, LOD and LOQ intensities by
# second method
lod_stp2 <- QDLOD_calib(data_rep_stp,LOB = TRUE)
# To visulaize all the results
lod_stp2
# To visualize only the LOB, LOD and LOQ intensities
slot(lod_stp2,"lobData")
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{Qdata}
