% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDLOD_exp.R
\docType{methods}
\name{QDLOD_exp}
\alias{QDLOD_exp}
\title{To compute the LOB, LOD and LOQ of sandwich assay}
\usage{
\S4method{QDLOD_exp}{Qdata}(Qdata, LOB = FALSE)
}
\arguments{
\item{Qdata}{object of S4 Class \code{Qdata}, which is an output of function
\code{Repsd}.}

\item{LOB}{Choose between TRUE and FALSE. FALSE is for computing the
LOD and LOQ intensities by first method, without involving the LOB. TRUE is for computing all
three LOB, LOD and LOQ intensities by second method, with involvement
of LOB in calculation of LOD too. Second method is more accurate method.}
}
\value{
An object of S4 Class \code{Qdata}, with a new slot of lobData to
save the results of LOB, LOD and LOQ of combined replicates of green
and red QDs.
}
\description{
Function \code{QDLOD_exp} works on the output of the function
\code{Repsd}. Function \code{QDLOD_exp} computes the limit of blank (LOB),
limit of detection (LOD) and limit of quantification (LOQ) intensities
of the sandwich assay. These intensities can be used to compute
the LOB, LOD and LOQ concentrations, based on the linear model equation.
}
\details{
Function \code{QDLOD_exp} works on the output of the function
\code{Repsd}. Function \code{QDLOD_exp} computes the limit of blank (LOB),
limit of detection (LOD) and limit of quantification (LOQ) intensities
of the sandwich assay. The results of LOB, LOD and LOQ intensities are
saved in the new slot lobData of the object of S4 Class \code{Qdata}. These intensities can be used to compute
the LOB, LOD and LOQ concentrations, based on the linear model equation.
}
\examples{
#To combine the replicates of sandwich assay
##To read the first file (.csv)
file_il_d_1 <- system.file("exData", "ImageJ_CRPi_IL6d_Qplex.csv", package = "Qplex")
data1 <- readQD(file = file_il_d_1, decp = ".")
## to read the second file (.csv)
file_il_i_1 <- system.file("exData", "ImageJ_CRPi_IL6i_Qplex.csv", package = "Qplex")
data2 <- readQD(file = file_il_i_1, decp = ".")
data <- ReadQD_combine(data1, data2)
data_rep_sand <- Repsd(data, calib = FALSE)
# 1) To compute the LOB, LOD and LOQ intensities by first
#method
lod_exp1 <- QDLOD_exp(data_rep_sand)
#To visualize all the results
lod_exp1
# To visualize the only the LOB, LOD and LOQ intensities
slot(lod_exp1,"lobData")
# 2) To compute the LOB, LOD and LOQ intensities by
#second method
lod_exp2 <- QDLOD_exp(data_rep_sand,LOB = TRUE)
#To visualize all the results
lod_exp2
# To visualize only the LOB, LOD and LOQ intensities
slot(lod_exp2,"lobData")
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{Qdata}
