% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qmodel_exp_15.R
\docType{methods}
\name{Qmodel_exp_15}
\alias{Qmodel_exp_15}
\title{To develop the statistical linear models and graphs with R-square
values and linear model equation.}
\usage{
\S4method{Qmodel_exp_15}{Qdata}(Qdata, greenQD = TRUE,
  graphtitle = "Green QD", W = 0.02, P = 0.02)
}
\arguments{
\item{Qdata}{object of S4 Class \code{Qdata}, which is an output of function
\code{Repsd}.}

\item{greenQD}{Choose between TRUE or FALSE. TRUE is for green QDs,
while FALSE is for red QDs. Default is TRUE.}

\item{graphtitle}{The title of the graph to be mentioned in the main of
the plot.}

\item{W}{a value to be provided for width. Default is 0.2}

\item{P}{a value to be provided for position_dodge. Default is 0.02}
}
\value{
Plot with R-square value, linear model equation and graph title.
}
\description{
Function \code{Qmodel_exp_15} works on the output of the function
\code{Repsd}. Function \code{Qmodel_exp_15} is designed for the sandwich assay.
It develops the statistical linear fit models of Normalized
Intensities vs Concentration, in a range with one value lesser than
the complete assay range. Here, it is in the range of 0-15 nM, after
removing the 20 nM concentration and respective intensity. It represents
the results in the form of plots, with embedded R-square value,
linear model equation and graph title in the plot.
}
\details{
Function \code{Qmodel_exp_15} works on the output of the function
\code{Repsd}. Function \code{Qmodel_exp_15} is designed for the sandwich assay.
It develops the statistical linear fit models of Normalized
Intensities vs Concentration, in a range with one value lesser than
the complete assay range. Here, it is in the range of 0-15 nM, after
removing the 20 nM concentration and respective intensity. It represents
the results in the form of plots, with embedded R-square value,
linear model equation and graph title in the plot.
}
\examples{
# Example is similar like Qmodel_exp
#'#To combine the replicates of sandwich assay
#To combine the replicates of sandwich assay
file_il_d_1 <- system.file("exData", "ImageJ_CRPi_IL6d.csv", package = "Qplex")
data1 <- readQD(file = file_il_d_1)
file_il_i_1 <- system.file("exData", "ImageJ_CRPi_IL6i.csv", package = "Qplex")
data2 <- readQD(file = file_il_i_1)
data <- ReadQD_combine(data1, data2)
data_rep_sand <- Repsd(data, calib = FALSE)#'file_il_d_1 <- system.file("exData", "ImageJ_CRPi_IL6d.csv", package = "Qplex")
data1 <- readQD(file = file_il_d_1)
file_il_i_1 <- system.file("exData", "ImageJ_CRPi_IL6i.csv", package = "Qplex")
data2 <- readQD(file = file_il_i_1)
data <- ReadQD_combine(data1, data2)
data_rep_sand <- Repsd(data, calib = FALSE)
# 1) linear fitting model for green QD
plot_sand_g <- Qmodel_exp(data_rep_sand)
# to visualize the plot
plot_sand_g
# 2) linear fitting model for red QD
plot_sand_r <- Qmodel_exp(data_rep_sand,
graphtitle = "Red QD")
# to visualize the plot
plot_sand_r
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{Qdata}
