% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QplexIntensity.R
\name{QplexIntensity}
\alias{QplexIntensity}
\title{To provide the intensity of the test line and control line, after
background processing.}
\usage{
QplexIntensity(image, fp = TRUE, cut1 = 1, cut2 = 1, icut = 0.2)
}
\arguments{
\item{image}{Image to be read in.}

\item{fp}{Choose between TRUE or FALSE. TRUE is for intensity of the
first peak, which is the intensity of the control line, while FALSE is
for intensity of the second peak, which is the intensity of the test
line.}

\item{cut1}{The first cut on the x-axis (index), from where the peak starts.}

\item{cut2}{The second cut on the x-axis (index), where the peak ends.}

\item{icut}{The intensity for background propagation to remove noise.}
}
\value{
The intensity for test line and control line after background
processing.
}
\description{
Function \code{QplexIntensity} provides the intensity
of the control line and test line after background processing.
}
\details{
#'@name QplexIntensity

Allows the user to compute the intensity of the control
line and test line after background processing. Before applying function
\code{QplexIntensity}, the user should apply the functions \code{readIQplexplot}
to exactly know the starting index values and intensity value to process
the signal properly by the function \code{readIQcutprocess}. The values of
the parameters used while applying the function \code{readIQcutprocess} helps
to exactly know the intensity for background processing, index cuts for
control line and test line signal. These values of parameters of function
\code{readIQcutprocess} are used in function \code{QplexIntensity} to get the
intensity after proper background propagation.
}
\examples{
# To compute the intensity of first peak (control line)
QplexIntensity(image  = image_test, fp = TRUE,
cut1 =1140, cut2 = 1200, icut = 0.004)
To compute the intensity of second peak (test line)
QplexIntensity(image  = image_test, fp = FALSE,
cut1 =1290, cut2 = 1340, icut = 0.0035)
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{Intensity}
