% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIQcutprocess.R
\name{readIQcutprocess}
\alias{readIQcutprocess}
\title{To cut the test line and control line peaks, along with
background processing.}
\usage{
readIQcutprocess(image, fp = TRUE, cut1 = 1, cut2 = 1, icut = 0.02)
}
\arguments{
\item{image}{Image to be read in.}

\item{fp}{Choose between TRUE or FALSE. TRUE is for the first peak
of the signal, which is the peak of control line, while FALSE is for
the second peak of the signal, which is the peak of the test line.}

\item{cut1}{The first cut on the x-axis (index), from where the peak starts.}

\item{cut2}{The second cut on the x-axis (index), where the peak ends.}

\item{icut}{The intensity for background propagation to remove error.}
}
\value{
The separated and background processed plots of the peaks of
control line and test line.
}
\description{
Function \code{readIQcutprocess} represents the separated
peaks of control line and test line in the form of the plot of intensity
vs index, after background processing.
}
\details{
Allows the user to represent the separated peaks of control line
and test line in the form of the plot of intensity vs index, after
background processing. Before applying function \code{readIQcutprocess},
the user should apply the function \code{readIQplexplot} to exactly
know the starting index values and intensity value to process the
signal properly by the function \code{readIQcutprocess}. It helps
to exactly know the intensity for background processing, index cuts
for control line and test line signal. These values of parameters
are used in the QplexIntensity to get the intensity values of the
control line and test line.
}
\examples{
#Attach the file through system.file and then apply the function
# To see the signal (plot) of first peak (control line)
#separately
readIQcutprocess(image  = image_test, fp = TRUE,
cut1 =1140, cut2 = 1200, icut = 0.004)
# To see the signal (plot) of second peak (test line)
# separately
readIQcutprocess(image  = image_test, fp = FALSE,
cut1 =1290, cut2 = 1340, icut = 0.0035)
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{Signal}
