% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIQplexplot.R
\name{readIQplexplot}
\alias{readIQplexplot}
\title{To display image or to display image data in the form of a signal
(plot of Intensity vs Index)}
\usage{
readIQplexplot(Image, plot = "No")
}
\arguments{
\item{Image}{Image to be read in}
}
\value{
image signal in the form of plot of Intensity vs Index.
}
\description{
Function \code{readIQplexplot} either displays the image or image
data in the form of a signal, which is a plot of intensity vs concentration
}
\details{
Allows the user either to read in the image and display
the image data in the form of a signal (a plot of intensity vs index).
The signal plot helps the user to understand that where the exact
signals of test line and cross line are present in the plot to
perform efficient background processing.
}
\examples{
# To read in the image and display in the form of signal plot
#attach the file of image through system.file and then read file.
#for example, let's say, the file attached is image1, then, to
# read image1 and show the plot code is as follows:
image_plot <- readIQplexplot(Image = image1)
# to display the image data in the form of plot (signal)
image_plot
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{Image}
