% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadQD_combine.R
\docType{methods}
\name{ReadQD_combine}
\alias{ReadQD_combine}
\title{To combine the raw intensity data files of sandwich lateral
flow assay}
\usage{
\S4method{ReadQD_combine}{Qdata}(Qdata, Qdata2)
}
\arguments{
\item{Qdata}{object of S4 Class \code{Qdata}. The first file read by
function \code{readQD} ans saved in the form of object of Class \code{Qdata}}

\item{Qdata2}{object of S4 Class \code{Qdata}. The second file read by
function \code{readQD} ans saved in the form of object of Class \code{Qdata}}
}
\value{
object of Class \code{"Qdata"} and save the resulting data in slot initialData
}
\description{
Function \code{ReadQD_combine} combines the raw intensity
files, read by function \code{readQD} and saved in the object of S4 class
\code{Qdata}. After combining it again save the combined data in the
form of an object of S4 Class \code{Qdata}.
}
\details{
Allows the user to combine the two raw intensity data files of sandwich assay,
read by function \code{readQD} and saved in the form of object of S4 class
\code{Qdata}. The output of function \code{ReadQD_combine} is also saved in the
form of object of S4 Class \code{Qdata}.
}
\examples{
##To read the first file (.csv)
file_il_d_1 <- system.file("exData", "ImageJ_CRPi_IL6d_Qplex.csv", package = "Qplex")
data1 <- readQD(file = file_il_d_1, decp = ".")
data1
## to read the second file (.csv)
file_il_i_1 <- system.file("exData", "ImageJ_CRPi_IL6i_Qplex.csv", package = "Qplex")
data2 <- readQD(file = file_il_i_1, decp = ".")
## to combine the data1 and data2
data <- ReadQD_combine(data1, data2)
##to visualise all the data after combining together
data
# to visualize the combined data
slot(data,"initialData")
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{Qdata}
