% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQplexImage.R
\name{readQplexImage}
\alias{readQplexImage}
\title{To display image}
\usage{
readQplexImage(Image)
}
\arguments{
\item{Image}{Image to be read in}
}
\value{
display image read in
}
\description{
Function \code{readQplexImage} displays the image read in
}
\details{
Allows the user to display the image
}
\examples{
# To read in the image and display the image
#attach the file of image through system.file, if it is not in
#same directory and then read file.
#for example, let's say, the file attached is image1, then, to
# read image1 and display image1, the code is as follows:
image_dis <- readQplexImage(Image = image1)
# to display the image data in the form of plot (signal)
image_dis
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{Image}
