% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Repsd.R
\docType{methods}
\name{Repsd}
\alias{Repsd}
\title{To combine the replicates, after normalisation of the intensity and compute the
standard deviations within replicates after normalisation of the
intensities}
\usage{
\S4method{Repsd}{Qdata}(Qdata, calib = TRUE)
}
\arguments{
\item{Qdata}{object of S4 Class \code{Qdata}, which is an output of function
\code{readQD} in case of streptavidin assay and an output of the function
\code{ReadQD_combine} in case of sandwich assay.}

\item{calib}{choose between TRUE or FALSE. TRUE is for streptavidin assay
and FALSE is for sandwich assay. Default is TRUE.}
}
\value{
object of Class \code{"Qdata"} and save the results in different slots, which are
described under section details.
}
\description{
Function \code{Repsd} works on the output of the function
\code{readQD} in case of streptavidin assay. In case of sandwich assay,
function \code{Repsd} works on the output of the function \code{ReadQD_combine}. Function
\code{Repsd} combines replicates after normlaization of the intensity
and computes the standard deviations within replicates after normalization
of the intensities. It saves the resulting data in the form of the object of
S4 Class \code{Qdata}.
}
\details{
Allows the user to combine the replicates after normalization of
the intensity and computes the standard deviation within replicates
after normalization of the intensities. The results are saved in different
slots of the Class \code{Qdata}. The common slots for the streptavidin assay
and sandwich assay for the output of the function \code{Repsd}:
\code{nsgData} : normalized intensities of green QDs
\code{nsrData} : normalized intensities of red QDs
\code{cgData} : combined replicates of normalized intensities of green QDs
\code{sdgData} : standard deviations of normalized intensities of replicates
of green QDs
\code{crData} : combined replicates of normalized intensities of red QDs
\code{sdrData} : standard deviations of normalized intensities of replicates
of red QDs
function Repsd separates the concentration slots also. The concentration
slots for streptavidin and sandwich assays are unique.
Slots which are unique for streptavidin assay:
\code{Concentration} : The concentration of green and red QDs, used in the assay
Slots which are unique for sandwich assay:
\code{Concentration_CRP} : Concentration used in case of green QDs
\code{Concentration_IL6} : COncentration used in case of red QDs
}
\examples{
##1) For streptavidin assay
file_cal_1 <- system.file("exData", "QD_calibration_Qplex_new.csv", package = "ReadqPCR")
data_stp <- readQD(file = file_cal_1, type = ".csv", decp = ",")
#to combine the replicates
data_rep_stp <- Repsd(data_stp)
#to visualise all resulting data
data_rep_stp
#to visualize the normalized green QD data
slot(data_rep_stp,"nsgData")
#to visualize the normalized red QD data
slot(data_rep_stp,"nsrData")
#to visualize the combined replicates of green QD data
slot(data_rep_stp,"cgData")
#to visualize the standard deviation of the green QD data
slot(data_rep_stp,"sdgData")
#to visualize the combined replicates of red QD data
slot(data_rep_stp,"crData")
#to visualize the standard deviation of the red QD data
slot(data_rep_stp,"sdrData")
## 2) for sandwich assay
##To read the first file (.csv)
file_il_d_1 <- system.file("exData", "ImageJ_CRPi_IL6d_Qplex.csv", package = "Qplex")
data1 <- readQD(file = file_il_d_1, decp = ".")
## to read the second file (.csv)
file_il_i_1 <- system.file("exData", "ImageJ_CRPi_IL6i_Qplex.csv", package = "Qplex")
data2 <- readQD(file = file_il_i_1, decp = ".")
## to combine the data1 and data2
data <- ReadQD_combine(data1, data2)
#to combine the replicates
data_rep_sand <- Repsd(data, calib = FALSE)
#to visualize all the data
data_rep_sand
#to visualize the normalized green QD data
slot(data_rep_sand,"nsgData")
#to visualize the normalized red QD data
slot(data_rep_sand,"nsrData")
#to visualize the combined replicates of green QD data
slot(data_rep_sand,"cgData")
#to visualize the standard deviation of the green QD data
slot(data_rep_sand,"sdgData")
#to visualize the combined replicates of red QD data
slot(data_rep_sand,"crData")
#to visualize the standard deviation of the red QD data
slot(data_rep_sand,"sdrData")
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{Qdata}
